requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
decorator==5.1.0
Django==3.1.5
django-environ==0.4.5
djangorestframework==3.12.4
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2020.5
sqlparse==0.4.1
traitlets==5.1.0
wcwidth==0.2.5
